+++ src/index.ts
@@ -5,7 +5,8 @@
  * @ version: 2020-07-06 11:21:37
  */
 export * from "./Application";
+export * from "./Component";
+export * from "./Context";
 export * from "./IApplication";
 export * from "./IContext";
 export * from "./Metadata";
-export * from "./Context";